requirements-dev.txt
# Testing Dependencies
pytest==9.0.1
pytest-asyncio==1.3.0
pytest-cov==7.0.0
pytest-env==1.2.0
pytest-mock==3.15.1

# Test Data Generation
# faker==38.2.0
factory-boy==3.3.3 # Faker is installed by factory-boy

# HTTP Testing
httpx==0.28.1
requests-mock==1.12.1

# Database Testing
pytest-postgresql==7.0.2
testing.postgresql==1.3.0


# Coverage Reporting
coverage[toml]==7.12.0
